+++ src/org/newdawn/slick/util/pathfinding/heuristics/ManhattanHeuristic.java
@@ -26,7 +26,7 @@ public class ManhattanHeuristic implements AStarHeuristic {
 	/**
 	 * @see AStarHeuristic#getCost(TileBasedMap, Mover, int, int, int, int)
 	 */
-	public int getCost(TileBasedMap map, Mover mover, int x, int y, int tx,
+	public float getCost(TileBasedMap map, Mover mover, int x, int y, int tx,
 			int ty) {
 		return minimumCost * (Math.abs(x-tx) + Math.abs(y-ty));
 	}
